pow blockchain
This commit introduces a minimal proof-of-work blockchain simulating
nodes that participate in block mining, broadcasting and validation.

In Python the GIL prevents 2 threads from running on separate cores
simultaneously but rather forces them to take turns. To make the mining
simulation more realistic, a separate process is spawned for each block
being mined. This in turn required a slightly more complex design to
account for the limitation that you can't pass all python objects to
other processes.

The pseudo-network is built with a net of MPSC (multi-producer,
single-consumer) queues. The building blocks primarly come from the
`multiprocessing` library.

Nodes perform proof-of-work mining by iterating over a nonce until a
block hash meets the difficulty requirement. Once mined, the block is
broadcasted to the common network, received again by the node and then
re-validated. This design choice was made to simplify synchronization
logic.